float en int wallah c'pas beau

+++ RType/Client/Includes/Sound.hpp
@@ -5,7 +5,7 @@
 // Login   <Volto@epitech.net>
 // 
 // Started on  Fri Nov 20 13:27:06 2015 Probola
-// Last update Tue Dec 15 15:00:00 2015 Probola
+// Last update Tue Dec 15 17:57:49 2015 Probola
 //
 
 #ifndef					_SOUND_HPP_
@@ -24,7 +24,7 @@ private:
 public:
   Sound();
   ~Sound();
-  bool					initializeSound(const std::string &, bool, int, int);
+  bool					initializeSound(const std::string &, bool, float, float);
   bool					initializeMusic(const std::string &, bool);
   void					playSound();
   void					playMusic();
